+++ docs/lib/mdx/index.ts
@@ -1,0 +1,9 @@
+// Templates
+export {
+	getTemplate,
+	getTemplateBreadcrumbs,
+	getTemplateSlugs,
+	getTemplateList,
+	getTemplatesBreadcrumbs,
+} from './templates';
+export type { Template, TemplateList } from './templates';
